+++ src/org/ods/orchestration/BuildStage.groovy
@@ -140,7 +140,7 @@ class BuildStage extends Stage {
         def index = 1
         def sanitizedRepos = failedRepos.collect { it ->
             (index++) + ".\tRepository id: " + it.id +
-            "\n\tBranch: " + it.branch + "\n\tRepository type: " + it.type }
+            "\n\tBranch: " + it.defaultBranch + "\n\tRepository type: " + it.type }
             .join("\n\n")
         return sanitizedRepos
     }